Found and fixed a bug in the GenerateUniqueGhostVariable method of the Name class

+++ Components/KnowledgeBase/WellFormedNames/ComposedName.cs
@@ -135,7 +135,7 @@ namespace KnowledgeBase.WellFormedNames
 
 
 
-		public override Name ReplaceUnboundVariables(int variableID)
+		public override Name ReplaceUnboundVariables(long variableID)
 		{
 			if (IsGrounded)
 				return (Name)Clone();
